sdk/aptos: consolidate utils
