+++ sdk/js/src/utils/aptos.ts
@@ -1,6 +1,5 @@
-import { ethers } from "ethers";
 import { sha3_256 } from "js-sha3";
-import { ChainId } from "../utils";
+import { ChainId, hex } from "../utils";
 
 export const deriveWrappedAssetAddress = (
   tokenBridgeAddress: string, // 32 bytes
@@ -12,8 +11,4 @@ export const deriveWrappedAssetAddress = (
   return sha3_256(
     Buffer.concat([hex(tokenBridgeAddress), chain, Buffer.from("::", "ascii"), hex(originAddress)]),
   );
-};
-
-export const hex = (x: string): Buffer => {
-  return Buffer.from(ethers.utils.hexlify(x, { allowMissingPrefix: true }).substring(2), "hex");
 };
