+++ sdk/js/src/utils/array.ts
@@ -1,10 +1,11 @@
 import { arrayify, zeroPad } from "@ethersproject/bytes";
 import { PublicKey } from "@solana/web3.js";
+import { ethers } from "ethers";
 import { hexValue, hexZeroPad, stripZeros } from "ethers/lib/utils";
 import {
   hexToNativeAssetStringAlgorand,
   nativeStringToHexAlgorand,
-  uint8ArrayToNativeStringAlgorand,
+  uint8ArrayToNativeStringAlgorand
 } from "../algorand";
 import { canonicalAddress, humanAddress } from "../cosmos";
 import { buildTokenId } from "../cosmwasm/address";
@@ -13,20 +14,20 @@ import {
   ChainId,
   ChainName,
   CHAIN_ID_ALGORAND,
-  CHAIN_ID_NEAR,
-  CHAIN_ID_INJECTIVE,
-  CHAIN_ID_OSMOSIS,
-  CHAIN_ID_SUI,
   CHAIN_ID_APTOS,
+  CHAIN_ID_INJECTIVE,
+  CHAIN_ID_NEAR,
+  CHAIN_ID_OSMOSIS,
+  CHAIN_ID_PYTHNET,
   CHAIN_ID_SOLANA,
+  CHAIN_ID_SUI,
   CHAIN_ID_TERRA,
   CHAIN_ID_TERRA2,
-  CHAIN_ID_WORMHOLE_CHAIN,
   CHAIN_ID_UNSET,
+  CHAIN_ID_WORMHOLE_CHAIN,
   coalesceChainId,
   isEVMChain,
-  isTerraChain,
-  CHAIN_ID_PYTHNET,
+  isTerraChain
 } from "./consts";
 
 /**
@@ -290,3 +291,7 @@ export function textToHexString(name: string): string {
 export function textToUint8Array(name: string): Uint8Array {
   return new Uint8Array(Buffer.from(name, "binary"));
 }
+
+export const hex = (x: string): Buffer => {
+  return Buffer.from(ethers.utils.hexlify(x, { allowMissingPrefix: true }).substring(2), "hex");
+};
