+++ sdk/js/src/utils/index.ts
@@ -1,6 +1,7 @@
-export * from "./consts";
-export * from "./createNonce";
-export * from "./parseVaa";
+export * from "./aptos";
 export * from "./array";
 export * from "./bigint";
+export * from "./consts";
+export * from "./createNonce";
 export * from "./near";
+export * from "./parseVaa";
